Renames export_thermal_model to export_lab_thermal_model and adds build_lab_thermal_model
